split up context/contextbuilder files
