+++ dodiscover/__init__.py
@@ -6,4 +6,4 @@ DoDiscover - a library for Python-based Causal Discovery
 from . import ci  # noqa: F401
 from ._protocol import EquivalenceClassProtocol, GraphProtocol
 from ._version import __version__  # noqa: F401
-from .context import context_builder
+from .context_builder import context_builder
